+++ packages/common/web/src/types/editorTypes.ts
@@ -7,6 +7,7 @@ import {
   ImageComponentData,
   VideoComponentData,
   FileComponentData,
+  HEADER_BLOCK,
 } from 'ricos-content';
 export interface Helpers extends BICallbacksForHelpers {
   openModal?: (modalProps: Record<string, unknown>) => void;
@@ -32,8 +33,14 @@ export interface Helpers extends BICallbacksForHelpers {
 }
 interface BICallbacksForHelpers extends BICallbacks {
   // makes version optional
-  onPluginAdd?(pluginId: string, entryPoint: string, version?: string): void;
-  onPluginAddSuccess?(pluginId: string, entryPoint: string, params, version?: string): void;
+  onPluginAdd?(pluginId: string, entryPoint: string, version?: string, contentId?: string): void;
+  onPluginAddSuccess?(
+    pluginId: string,
+    entryPoint: string,
+    params,
+    version?: string,
+    contentId?: string
+  ): void;
   isPreview?: () => boolean;
   onPluginAction?: OnPluginAction;
 }
@@ -55,4 +62,14 @@ export const DROPDOWN_OPTIONS_TO_DOC_STYLE_TYPE = {
   H4: 'headerFour',
   H5: 'headerFive',
   H6: 'headerSix',
+};
+
+export const DRAFT_TO_RICOS_DOC_TYPE = {
+  [HEADER_BLOCK.ONE]: 'headerOne',
+  [HEADER_BLOCK.TWO]: 'headerTwo',
+  [HEADER_BLOCK.THREE]: 'headerThree',
+  [HEADER_BLOCK.FOUR]: 'headerFour',
+  [HEADER_BLOCK.FIVE]: 'headerFive',
+  [HEADER_BLOCK.SIX]: 'headerSix',
+  [HEADER_BLOCK.PARAGRAPH]: 'paragraph',
 };