+++ packages/toolbars-new/web/src/modals/heading/HeadingsPanel.jsx
@@ -1,64 +1,122 @@
 /* eslint-disable react/prop-types */
 import React, { Component } from 'react';
-import styles from '../panels/styles.scss';
+import styles from './customPanelStyles.scss';
 import MobilePanel from '../panels/MobilePanel';
 import DesktopPanel from '../panels/DesktopPanel';
+import UpdateHeadingPanel from './UpdateHeadingPanel';
 import classNames from 'classnames';
-import { mergeStyles, GlobalContext } from 'wix-rich-content-common';
+import {
+  mergeStyles,
+  GlobalContext,
+  DROPDOWN_OPTIONS_TO_DOC_STYLE_TYPE,
+} from 'wix-rich-content-common';
 import { HEADER_TYPE_MAP } from 'wix-rich-content-plugin-commons';
+import { getFontSizeNumber, hasStyleChanges } from '../../Toolbar/utils';
+import { omit, cloneDeep, pick, isEmpty } from 'lodash';
+import { ColorsIcon } from './icons';
 
-export const DEFAULT_HEADERS_DROPDOWN_OPTIONS = Object.freeze([
-  'P',
-  'H1',
-  'H2',
-  'H3',
-  'H4',
-  'H5',
-  'H6',
-]);
+const DEFAULT_HEADERS_DROPDOWN_OPTIONS = Object.freeze(['H1', 'H2', 'H3', 'H4', 'H5', 'H6', 'P']);
 
 class HeadingsPanel extends Component {
   constructor(props) {
     super(props);
-    this.state = { heading: props.currentSelect };
+    this.state = { openOption: '' };
     this.styles = mergeStyles({ styles, theme: props.theme });
-  }
-
-  componentDidUpdate() {
-    const { heading } = this.state;
-    if (heading !== this.props.currentSelect) {
-      // eslint-disable-next-line react/no-did-update-set-state
-      this.setState({ heading: this.props.currentSelect });
-    }
   }
 
   static contextType = GlobalContext;
 
   onSaveHeading = (type, clickFromKeyboard) => {
     this.props?.onToolbarButtonClick?.(type);
-    return this.props.onSave({ clickFromKeyboard, data: type });
+    this.setState({ openOption: '' });
+    return this.props.onSave({ data: type, clickFromKeyboard });
   };
 
-  defaultHeadings = () => {
-    const defaults = DEFAULT_HEADERS_DROPDOWN_OPTIONS.map(heading => ({
-      text: this.props.translateHeading(heading, this.props.t),
-      commandKey: HEADER_TYPE_MAP[heading],
-    }));
-    return defaults;
+  onUpdateHeading = type => {
+    const { onToolbarButtonClick, documentStyle, currentInlineStyles } = this.props;
+    onToolbarButtonClick?.('Update ' + type);
+    this.setState({ openOption: '' });
+    const omitValues = Object.entries(currentInlineStyles)
+      .filter(([key, value]) => !value)
+      .map(([key, value]) => key);
+    return this.props.onChange({
+      [type]: omit({ ...documentStyle[type], ...currentInlineStyles }, omitValues),
+    });
+  };
+
+  onResetHeading = type => {
+    this.props?.onToolbarButtonClick?.('Reset ' + type);
+    this.setState({ openOption: '' });
+    return this.props.onChange({ [type]: {} });
+  };
+
+  getUpdatePanelModal = heading => {
+    const { t, currentSelect, currentInlineStyles, documentStyle } = this.props;
+    const documentStyleType = DROPDOWN_OPTIONS_TO_DOC_STYLE_TYPE[heading];
+    const headerType = HEADER_TYPE_MAP[heading];
+    return (
+      <UpdateHeadingPanel
+        optionName={this.props.translateHeading(heading, t)}
+        onApply={clickFromKeyboard => this.onSaveHeading(headerType, clickFromKeyboard)}
+        onReset={() => this.onResetHeading(documentStyleType)}
+        onUpdate={() => this.onUpdateHeading(documentStyleType)}
+        t={this.props.t}
+        resetEnabled={() => !isEmpty(this.props.documentStyle[documentStyleType])}
+        updateEnabled={() =>
+          headerType === currentSelect &&
+          hasStyleChanges(currentSelect, currentInlineStyles, documentStyle)
+        }
+      />
+    );
+  };
+
+  getHeadingsOptions = () => {
+    return (this.props.customHeadings || DEFAULT_HEADERS_DROPDOWN_OPTIONS).map(heading => {
+      let customizeOptions = {};
+      if (this.props.allowHeadingCustomization) {
+        const { documentStyle, wiredFontStyles } = this.props;
+        const documentStyleType = DROPDOWN_OPTIONS_TO_DOC_STYLE_TYPE[heading];
+        const optionWiredStyles = wiredFontStyles[documentStyleType];
+        const subText = optionWiredStyles['font-size'];
+        const fontSize = parseInt(getFontSizeNumber(subText)) > 20 ? '20px' : subText;
+        const fontFamily = optionWiredStyles['font-family'];
+        customizeOptions = {
+          subText,
+          style: { fontSize, fontFamily },
+          modal: !this.props.isMobile && this.getUpdatePanelModal(heading),
+          icon: () => (
+            <ColorsIcon
+              colors={pick(documentStyle[documentStyleType], ['color', 'background-color'])}
+            />
+          ),
+        };
+      }
+      return {
+        text: this.props.translateHeading(heading, this.props.t),
+        tooltip: this.props.translateHeading(heading, this.props.t, true),
+        commandKey: HEADER_TYPE_MAP[heading],
+        ...customizeOptions,
+      };
+    });
+  };
+
+  onUpdateModalOpen = type => {
+    this.setState({ openOption: type });
   };
 
   render() {
-    const { isMobile, t, currentSelect } = this.props;
-    const panelHeader = t('Headings');
+    const { isMobile, t, currentSelect, onCancel, allowHeadingCustomization } = this.props;
+    const { openOption } = this.state;
+    const panelHeader = t('FormattingToolbar_HeadingsPanelHeader');
 
     const panel = isMobile ? (
       <MobilePanel
         {...{
           currentSelect,
           panelHeader,
-          options: this.defaultHeadings(),
+          options: this.getHeadingsOptions(),
           onChange: this.onSaveHeading,
-
+          onCancel,
           t,
         }}
       />
@@ -66,13 +124,20 @@ class HeadingsPanel extends Component {
       <DesktopPanel
         {...{
           currentSelect,
-          options: this.defaultHeadings(),
-          onChange: this.onSaveHeading,
+          options: this.getHeadingsOptions(),
+          customPanelOptions: allowHeadingCustomization && {
+            openOption,
+            inline: true,
+            onOpen: this.onUpdateModalOpen,
+          },
           t,
+          onChange: this.onSaveHeading,
+          displayIconAndText: allowHeadingCustomization,
         }}
       />
     );
     return (
+      // eslint-disable-next-line jsx-a11y/no-static-element-interactions
       <div
         className={classNames(styles.panel_Container, {
           [styles.mobile_Container]: isMobile,